Initial work: Move helper functions into features/Linodes/utilities.ts & LinodeNetworking/utils.ts; display VPC IPv6 address in Summary panel
