+++ packages/manager/src/features/Linodes/LinodeEntityDetailBody.tsx
@@ -1,5 +1,5 @@
 import { usePreferences, useProfile } from '@linode/queries';
-import { Box, Typography } from '@linode/ui';
+import { Box, Stack, Typography } from '@linode/ui';
 import { pluralize } from '@linode/utilities';
 import { useMediaQuery } from '@mui/material';
 import Grid from '@mui/material/Grid';
@@ -13,6 +13,8 @@ import { useIsDiskEncryptionFeatureEnabled } from 'src/components/Encryption/uti
 import { Link } from 'src/components/Link';
 import { useKubernetesBetaEndpoint } from 'src/features/Kubernetes/kubeUtils';
 import { AccessTable } from 'src/features/Linodes/AccessTable';
+import { ipTableId } from 'src/features/Linodes/LinodesDetail/LinodeNetworking/utils';
+import { useFlags } from 'src/hooks/useFlags';
 import { useKubernetesClusterQuery } from 'src/queries/kubernetes';
 
 import { EncryptedStatus } from '../Kubernetes/KubernetesClusterDetail/NodePoolsDisplay/NodeTable';
@@ -22,9 +24,9 @@ import {
   StyledBodyGrid,
   StyledColumnLabelGrid,
   StyledCopyTooltip,
-  StyledIPv4Box,
-  StyledIPv4Item,
-  StyledIPv4Label,
+  StyledIPBox,
+  StyledIPItem,
+  StyledIPLabel,
   StyledLabelBox,
   StyledListItem,
   StyledVPCBox,
@@ -32,8 +34,8 @@ import {
 } from './LinodeEntityDetail.styles';
 import { LinodeEntityDetailRowConfigFirewall } from './LinodeEntityDetailRowConfigFirewall';
 import { LinodeEntityDetailRowInterfaceFirewall } from './LinodeEntityDetailRowInterfaceFirewall';
-import { ipTableId } from './LinodesDetail/LinodeNetworking/LinodeIPAddresses';
 import { lishLink, sshLink } from './LinodesDetail/utilities';
+import { getSubnetsString, getVPCIPv4, getVPCIPv6 } from './utilities';
 
 import type { LinodeHandlers } from './LinodesLanding/LinodesLanding';
 import type {
@@ -43,7 +45,6 @@ import type {
   Linode,
   LinodeCapabilities,
   LinodeInterface,
-  Subnet,
   VPC,
 } from '@linode/api-v4';
 import type { TypographyProps } from '@linode/ui';
@@ -107,6 +108,8 @@ export const LinodeEntityDetailBody = React.memo((props: BodyProps) => {
     vpcLinodeIsAssignedTo,
   } = props;
 
+  const flags = useFlags();
+
   const { data: profile } = useProfile();
 
   const { data: maskSensitiveDataPreference } = usePreferences(
@@ -121,6 +124,7 @@ export const LinodeEntityDetailBody = React.memo((props: BodyProps) => {
 
   const isLinodeInterface = interfaceGeneration === 'linode';
   const vpcIPv4 = getVPCIPv4(interfaceWithVPC);
+  const vpcIPv6 = getVPCIPv6(interfaceWithVPC);
 
   // @TODO LDE: Remove usages of this variable once LDE is fully rolled out (being used to determine formatting adjustments currently)
   const isDisplayingEncryptedStatus =
@@ -343,45 +347,87 @@ export const LinodeEntityDetailBody = React.memo((props: BodyProps) => {
               alignItems: 'center',
               margin: 0,
               padding: '0 0 8px 0',
-              [theme.breakpoints.down('md')]: {
+              [theme.breakpoints.down('lg')]: {
                 alignItems: 'start',
                 display: 'flex',
                 flexDirection: 'column',
               },
             }}
           >
-            <StyledVPCBox>
-              <StyledListItem sx={{ paddingLeft: 0 }}>
-                <StyledLabelBox component="span">Label:</StyledLabelBox>{' '}
-                <Link
-                  data-testid="assigned-vpc-label"
-                  to={`/vpcs/${vpcLinodeIsAssignedTo.id}`}
-                >
-                  {vpcLinodeIsAssignedTo.label}
-                </Link>
-              </StyledListItem>
-            </StyledVPCBox>
-            <StyledVPCBox>
-              <StyledListItem sx={{ ...sxLastListItem }}>
-                <StyledLabelBox component="span" data-testid="subnets-string">
-                  Subnet:
-                </StyledLabelBox>{' '}
-                {getSubnetsString(linodeAssociatedSubnets ?? [])}
-              </StyledListItem>
-            </StyledVPCBox>
-            {vpcIPv4 && (
-              <StyledIPv4Box>
-                <StyledIPv4Label data-testid="vpc-ipv4">
-                  VPC IPv4
-                </StyledIPv4Label>
-                <StyledIPv4Item component="span" data-testid="vpc-ipv4">
-                  <CopyTooltip copyableText text={vpcIPv4} />
-                  <Box sx={{ ml: 1, position: 'relative', top: 1 }}>
-                    <StyledCopyTooltip text={vpcIPv4} />
-                  </Box>
-                </StyledIPv4Item>
-              </StyledIPv4Box>
-            )}
+            <Stack
+              direction="row"
+              sx={{
+                [theme.breakpoints.down('md')]: {
+                  alignItems: 'start',
+                  display: 'flex',
+                  flexDirection: 'column',
+                },
+              }}
+            >
+              <StyledVPCBox>
+                <StyledListItem sx={{ paddingLeft: 0 }}>
+                  <StyledLabelBox component="span">Label:</StyledLabelBox>{' '}
+                  <Link
+                    data-testid="assigned-vpc-label"
+                    to={`/vpcs/${vpcLinodeIsAssignedTo.id}`}
+                  >
+                    {vpcLinodeIsAssignedTo.label}
+                  </Link>
+                </StyledListItem>
+              </StyledVPCBox>
+              <StyledVPCBox>
+                <StyledListItem sx={{ ...sxLastListItem }}>
+                  <StyledLabelBox component="span" data-testid="subnets-string">
+                    Subnet:
+                  </StyledLabelBox>{' '}
+                  {getSubnetsString(linodeAssociatedSubnets ?? [])}
+                </StyledListItem>
+              </StyledVPCBox>
+            </Stack>
+            <Stack
+              direction="row"
+              sx={{
+                [theme.breakpoints.down('md')]: {
+                  alignItems: 'start',
+                  display: 'flex',
+                  flexDirection: 'column',
+                  '& > div:first-child': {
+                    marginBottom: theme.spacingFunction(2),
+                  },
+                },
+                [theme.breakpoints.down(1275)]: {
+                  '& > div:first-child': {
+                    marginLeft: theme.spacingFunction(0),
+                  },
+                },
+              }}
+            >
+              {vpcIPv4 && (
+                <StyledIPBox>
+                  <StyledIPLabel data-testid="vpc-ipv4">VPC IPv4</StyledIPLabel>
+                  <StyledIPItem component="span" data-testid="vpc-ipv4">
+                    <CopyTooltip copyableText text={vpcIPv4} />
+                    <Box sx={{ ml: 1, position: 'relative', top: 1 }}>
+                      <StyledCopyTooltip text={vpcIPv4} />
+                    </Box>
+                  </StyledIPItem>
+                </StyledIPBox>
+              )}
+              {flags.vpcIpv6 &&
+                vpcIPv6 && ( // @TODO VPC IPv6: remove flag check once VPC IPv6 is fully rolled out
+                  <StyledIPBox>
+                    <StyledIPLabel data-testid="vpc-ipv6">
+                      VPC IPv6
+                    </StyledIPLabel>
+                    <StyledIPItem component="span" data-testid="vpc-ipv6">
+                      <CopyTooltip copyableText text={vpcIPv6} />
+                      <Box sx={{ ml: 1, position: 'relative', top: 1 }}>
+                        <StyledCopyTooltip text={vpcIPv6} />
+                      </Box>
+                    </StyledIPItem>
+                  </StyledIPBox>
+                )}
+            </Stack>
           </Grid>
         </Grid>
       )}
@@ -403,28 +449,3 @@ export const LinodeEntityDetailBody = React.memo((props: BodyProps) => {
     </>
   );
 });
-
-export const getSubnetsString = (data: Subnet[]) => {
-  const firstThreeSubnets = data.slice(0, 3);
-  const subnetLabels = firstThreeSubnets.map((subnet) => subnet.label);
-  const firstThreeSubnetsString = subnetLabels.join(', ');
-
-  return data.length > 3
-    ? firstThreeSubnetsString.concat(`, plus ${data.length - 3} more.`)
-    : firstThreeSubnetsString;
-};
-
-export const getVPCIPv4 = (
-  interfaceWithVPC: Interface | LinodeInterface | undefined
-) => {
-  if (interfaceWithVPC) {
-    if ('purpose' in interfaceWithVPC) {
-      return interfaceWithVPC.ipv4?.vpc;
-    }
-    return interfaceWithVPC.vpc?.ipv4?.addresses.find(
-      (address) => address.primary
-    )?.address;
-  }
-
-  return undefined;
-};
