+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeNetworking/LinodeIPAddressRow.test.tsx
@@ -4,13 +4,10 @@ import * as React from 'react';
 
 import { vpcIPFactory } from 'src/factories';
 import { PUBLIC_IP_ADDRESSES_CONFIG_INTERFACE_TOOLTIP_TEXT } from 'src/features/Linodes/constants';
-import {
-  createVPCIPv4Display,
-  ipResponseToDisplayRows,
-} from 'src/features/Linodes/LinodesDetail/LinodeNetworking/LinodeIPAddresses';
 import { renderWithTheme, wrapWithTableBody } from 'src/utilities/testHelpers';
 
 import { LinodeIPAddressRow } from './LinodeIPAddressRow';
+import { createVPCIPv4Display, ipResponseToDisplayRows } from './utils';
 
 import type { IPAddressRowHandlers } from './LinodeIPAddressRow';
