+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeNetworking/LinodeIPAddresses.test.tsx
@@ -1,7 +1,7 @@
 import { ipAddressFactory } from 'src/factories/networking';
 
-import { createType, ipResponseToDisplayRows } from './LinodeIPAddresses';
 import { listIPv6InRange } from './LinodeIPAddressRow';
+import { createType, ipResponseToDisplayRows } from './utils';
 
 import type { LinodeIPsResponse } from '@linode/api-v4/lib/linodes';
